backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0

# MongoDB async driver
pymongo==4.10.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.12

# Environment variables
python-dotenv==1.0.1

# Dedalus AI
dedalus-labs==0.0.1

# Pydantic for validation
pydantic==2.12.4
pydantic-core==2.41.5

# Backtesting tools
yfinance==0.2.66
backtesting==0.6.5
TA-Lib==0.6.8
pandas==2.3.3
numpy==2.3.4

# Additional dependencies for backtesting
bokeh==3.8.1
requests==2.32.5
beautifulsoup4==4.14.2

# Twitter API
python-twitter-v2==0.9.3
